backend/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
anthropic==0.42.0
firebase-admin==6.6.0
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
